+++ lib/src/brotli/stubs/decompress_filter.dart
@@ -16,11 +16,8 @@ class BrotliDecompressFilter extends DartCodecFilterBase {
   BrotliDecompressFilter(
       {bool ringBufferReallocation = true,
       bool largeWindow = false,
-      int inputBufferLength = 16386,
-      int outputBufferLength = 16386})
-      : super(
-            inputBufferLength: inputBufferLength,
-            outputBufferLength: outputBufferLength) {
+      super.inputBufferLength,
+      super.outputBufferLength}) {
     throw UnsupportedError('No CodecFilter Implementation');
   }